Extracted functions from a single source file and split them into multiple, named ones for easier navigation
